frontend/chat: regenerating LLM response should be immediately part of the thread

+++ src/packages/frontend/chat/actions.ts
@@ -758,15 +758,24 @@ export class ChatActions extends Actions<ChatState> {
 
     // The sender_id might change if we explicitly set the LLM model.
     if (tag === "regenerate" && llm != null) {
+      if (!this.store) return;
+      const messages = this.store.get("messages");
+      if (!messages) return;
       if (message.sender_id !== sender_id) {
         // if that happens, create a new message with the existing history and the new sender_id
         const cur = this.syncdb.get_one({ event: "chat", date });
+        if (cur == null) return;
+        const reply_to = getReplyToRoot(
+          cur.toJS() as any as ChatMessage,
+          messages,
+        );
         this.syncdb.delete({ event: "chat", date });
         this.syncdb.set({
           date,
           history: cur?.get("history") ?? [],
           event: "chat",
           sender_id,
+          reply_to,
         });
       }
     }
